9C-491: Renames some classes and objects
