+++ modules/api/src/main/scala/com.fortysevendeg.ninecards/api/JsonFormats.scala
@@ -5,7 +5,7 @@ import com.fortysevendeg.ninecards.api.messages.GooglePlayMessages._
 import com.fortysevendeg.ninecards.api.messages.InstallationsMessages._
 import com.fortysevendeg.ninecards.api.messages.SharedCollectionMessages._
 import com.fortysevendeg.ninecards.api.messages.UserMessages._
-import com.fortysevendeg.ninecards.processes.messages.GooglePlayMessages.CategorizedApp
+import com.fortysevendeg.ninecards.processes.messages.ApplicationMessages.CategorizedApp
 import io.circe.{ Decoder, Encoder, Json }
 import org.joda.time.DateTime
 import org.joda.time.format.DateTimeFormat
