+++ modules/processes/src/main/scala/com/fortysevendeg/ninecards/processes/ApplicationProcesses.scala
@@ -2,24 +2,24 @@ package com.fortysevendeg.ninecards.processes
 
 import cats.free.Free
 import com.fortysevendeg.ninecards.processes.converters.Converters._
-import com.fortysevendeg.ninecards.processes.messages.GooglePlayMessages._
+import com.fortysevendeg.ninecards.processes.messages.ApplicationMessages._
 import com.fortysevendeg.ninecards.services.free.algebra.GooglePlay
 
-class GooglePlayProcesses[F[_]](implicit googlePlayServices: GooglePlay.Services[F]) {
+class ApplicationProcesses[F[_]](implicit services: GooglePlay.Services[F]) {
 
   def categorizeApps(
     packagesName: List[String],
     authParams: AuthParams
   ): Free[F, CategorizeAppsResponse] =
-    googlePlayServices.resolveMany(
+    services.resolveMany(
       packageNames = packagesName,
       auth         = toAuthParamsServices(authParams)
     ) map toCategorizeAppsResponse
 }
 
-object GooglePlayProcesses {
+object ApplicationProcesses {
 
-  implicit def googlePlayProcesses[F[_]](implicit googlePlayServices: GooglePlay.Services[F]) =
-    new GooglePlayProcesses
+  implicit def applicationProcesses[F[_]](implicit services: GooglePlay.Services[F]) =
+    new ApplicationProcesses
 
 }
