+++ modules/processes/src/main/scala/com/fortysevendeg/ninecards/processes/converters/Converters.scala
@@ -2,7 +2,7 @@ package com.fortysevendeg.ninecards.processes.converters
 
 import java.sql.Timestamp
 
-import com.fortysevendeg.ninecards.processes.messages.GooglePlayMessages._
+import com.fortysevendeg.ninecards.processes.messages.ApplicationMessages._
 import com.fortysevendeg.ninecards.processes.messages.InstallationsMessages._
 import com.fortysevendeg.ninecards.processes.messages.SharedCollectionMessages._
 import com.fortysevendeg.ninecards.processes.messages.UserMessages.LoginResponse
