+++ modules/processes/src/test/scala/com/fortysevendeg/ninecards/processes/GooglePlayProcessesSpec.scala
@@ -3,7 +3,7 @@ package com.fortysevendeg.ninecards.processes
 import cats.free.Free
 import com.fortysevendeg.ninecards.processes.NineCardsServices.NineCardsServices
 import com.fortysevendeg.ninecards.processes.converters.Converters._
-import com.fortysevendeg.ninecards.processes.messages.GooglePlayMessages.AuthParams
+import com.fortysevendeg.ninecards.processes.messages.ApplicationMessages.AuthParams
 import com.fortysevendeg.ninecards.services.free.algebra.GooglePlay.Services
 import com.fortysevendeg.ninecards.services.free.domain.GooglePlay.{ AppInfo, AppsInfo }
 import org.specs2.matcher.Matchers
@@ -21,7 +21,7 @@ trait GooglePlayProcessesSpecification
   trait BasicScope extends Scope {
 
     implicit val googlePlayServices: Services[NineCardsServices] = mock[Services[NineCardsServices]]
-    implicit val googlePlayProcesses = new GooglePlayProcesses[NineCardsServices]
+    implicit val applicationProcesses = new ApplicationProcesses[NineCardsServices]
 
     googlePlayServices.resolveMany(Nil, toAuthParamsServices(authParams)) returns
       Free.pure(AppsInfo(Nil, Nil))
@@ -63,14 +63,14 @@ class GooglePlayProcessesSpec extends GooglePlayProcessesSpecification {
 
   "categorizeApps" should {
     "return empty items and errors lists if an empty list of apps is provided" in new BasicScope {
-      val response = googlePlayProcesses.categorizeApps(Nil, authParams).foldMap(testInterpreters)
+      val response = applicationProcesses.categorizeApps(Nil, authParams).foldMap(testInterpreters)
 
       response.errors should beEmpty
       response.items should beEmpty
     }
 
     "return items and errors lists for a non empty list of apps" in new BasicScope {
-      val response = googlePlayProcesses.categorizeApps(packageNames, authParams).foldMap(testInterpreters)
+      val response = applicationProcesses.categorizeApps(packageNames, authParams).foldMap(testInterpreters)
 
       response.errors shouldEqual missing
       forall(response.items) { item ⇒
